Moved files in to pkg, cmd, configs after review
