+++ cmd/onos-config-manager/config-manager.go
@@ -58,13 +58,13 @@ package main
 import (
 	"flag"
 	"fmt"
-	"github.com/opennetworkinglab/onos-config/events"
-	"github.com/opennetworkinglab/onos-config/listener"
-	"github.com/opennetworkinglab/onos-config/northbound/restconf"
-	"github.com/opennetworkinglab/onos-config/shell"
-	"github.com/opennetworkinglab/onos-config/southbound/topocache"
-	"github.com/opennetworkinglab/onos-config/store"
-	"github.com/opennetworkinglab/onos-config/southbound/synchronizer"
+	"github.com/opennetworkinglab/onos-config/cmd/onos-config-manager/shell"
+	"github.com/opennetworkinglab/onos-config/pkg/events"
+	"github.com/opennetworkinglab/onos-config/pkg/listener"
+	"github.com/opennetworkinglab/onos-config/pkg/northbound/restconf"
+	"github.com/opennetworkinglab/onos-config/pkg/southbound/synchronizer"
+	"github.com/opennetworkinglab/onos-config/pkg/southbound/topocache"
+	"github.com/opennetworkinglab/onos-config/pkg/store"
 	"log"
 	"log/syslog"
 	"os"
