+++ cmd/onos-config-manager/shell/cmdshell.go
@@ -18,10 +18,10 @@ package shell
 import (
 	"bufio"
 	"fmt"
-	"github.com/opennetworkinglab/onos-config/events"
-	"github.com/opennetworkinglab/onos-config/southbound/topocache"
-	"github.com/opennetworkinglab/onos-config/store"
-	"github.com/opennetworkinglab/onos-config/store/change"
+	"github.com/opennetworkinglab/onos-config/pkg/events"
+	"github.com/opennetworkinglab/onos-config/pkg/southbound/topocache"
+	"github.com/opennetworkinglab/onos-config/pkg/store"
+	"github.com/opennetworkinglab/onos-config/pkg/store/change"
 	"html/template"
 	"os"
 	"strconv"
