+++ pkg/northbound/restconf/restconf.go
@@ -21,8 +21,8 @@ import (
 	"crypto/sha1"
 	"encoding/json"
 	"fmt"
-	"github.com/opennetworkinglab/onos-config/listener"
-	"github.com/opennetworkinglab/onos-config/store"
+	"github.com/opennetworkinglab/onos-config/pkg/listener"
+	"github.com/opennetworkinglab/onos-config/pkg/store"
 	"html/template"
 	"io"
 	"log"
