+++ pkg/southbound/synchronizer/factory.go
@@ -15,10 +15,10 @@
 package synchronizer
 
 import (
-	"github.com/opennetworkinglab/onos-config/events"
-	"github.com/opennetworkinglab/onos-config/listener"
-	"github.com/opennetworkinglab/onos-config/southbound/topocache"
-	"github.com/opennetworkinglab/onos-config/store"
+	"github.com/opennetworkinglab/onos-config/pkg/events"
+	"github.com/opennetworkinglab/onos-config/pkg/listener"
+	"github.com/opennetworkinglab/onos-config/pkg/southbound/topocache"
+	"github.com/opennetworkinglab/onos-config/pkg/store"
 	"log"
 )
