+++ pkg/southbound/synchronizer/synchronizer.go
@@ -16,10 +16,10 @@
 package synchronizer
 
 import (
-	"github.com/opennetworkinglab/onos-config/events"
-	"github.com/opennetworkinglab/onos-config/southbound"
-	"github.com/opennetworkinglab/onos-config/southbound/topocache"
-	"github.com/opennetworkinglab/onos-config/store"
+	"github.com/opennetworkinglab/onos-config/pkg/events"
+	"github.com/opennetworkinglab/onos-config/pkg/southbound"
+	"github.com/opennetworkinglab/onos-config/pkg/southbound/topocache"
+	"github.com/opennetworkinglab/onos-config/pkg/store"
 	"log"
 )
